[FIX] cyc: submodule updated Sale Order PDF: amounts were not being exactly and Number to letters Before this commit:
1. The pdf fields were calculated using the value multiplied by the
exchange rate, in theory this will give the same value, but if
we calculate the exchange rate before calculate the taxes the final
value will no match by centavos.

2. A method for convert numbers in sale order to its form in letters
were implemented in a not standard way, also odoo already has its
own method already implemented in the currency model.

Now:

1. The taxes are calculated for the pdf after convert to USD, and
then the taxes and the order untaxed amount are sum to get the final
value.

2. We are using the method from currency model and the old
implementation was removed.

Extra: we take the opportunity to clean code.

Task#47742
Issue#15745
Task#48643