[REV] cyc: Fix adding freight, bond, and passing changes
Before this commit a refactor of some methods were applied at [1],
In this commit we are reverting the changes applied to the method that
computes the taxes, subtotal and total.

The method were rewritten incorrectly, now it has a better structure
and it fallows the original implementation.

A TODO were added, because the implementation of this feature is not the
best and it has to be reviewed in the future.

[1] https://git.vauxoo.com/vauxoo/grupo-cyc/-/commit/dc4127192c4e6e64f5c2e3c77c4d68f3e19383fe

Issue#17624